+++ src/schemathesis/contrib/__init__.py
@@ -1,0 +1,1 @@
+from . import unique_data